# Speed up queries for grade editor.