--- a/app/controllers/concerns/grade_editor.rb
+++ b/app/controllers/concerns/grade_editor.rb
@@ -8,11 +8,11 @@ module GradeEditor
   # @return [Grade, GradeComment] the grade or comment that corresponds to the
   #     attributes in the given parameters
   def find_or_build_feedback(klass, params)
-    @subject = AssignmentFeedback.find_subject params[:subject_type],
-                                               params[:subject_id]
-    @metric = current_course.assignment_metrics.find params[:metric_id]
-    unless feedback = klass.where(subject: @subject, metric: @metric).first
-      feedback = klass.new subject: @subject, metric: @metric
+    subject = AssignmentFeedback.find_subject params[:subject_type],
+                                              params[:subject_id]
+    metric = current_course.assignment_metrics.find params[:metric_id]
+    unless feedback = klass.find_by(subject: subject, metric: metric)
+      feedback = klass.new subject: subject, metric: metric
     end
     # TODO: check that the subject is reasonable and belongs to the course
     feedback.grader = current_user