--- a/app/helpers/grades_helper.rb
+++ b/app/helpers/grades_helper.rb
@@ -5,18 +5,6 @@ module GradesHelper
     options_from_collection_for_select assignments, :id, :name, selected.id
   end
 
-  # The score of the grade for the given metric and subject.
-  def grade_score_text(metric, subject)
-    grade = Grade.find_by metric: metric, subject: subject
-    grade && grade.score
-  end
-
-  # The text of the comment for the given metric and subject.
-  def grade_comment_text(metric, subject)
-    comment = GradeComment.find_by metric: metric, subject: subject
-    comment && comment.text
-  end
-
   # The weighted sum of all the scores in an array of grades.
   def grades_sum(grades)
     grades.map { |grade| grade ? grade.score * grade.metric.weight : 0 }.sum