--- a/app/views/grades/edit.html.erb
+++ b/app/views/grades/edit.html.erb
@@ -1,1 +1,2 @@
-<%= render 'grades/individual_form', metric: @metric, subject: @subject %>
+<%= render 'grades/individual_form', metric: @metric, grade: @grade,
+      comment: @comment %>