Loading the editor before: 1423ms (1260.6ms Views, 131.4ms ActiveRecord)
Loading the editor now: 589ms (532.2ms Views, 8.6ms ActiveRecord)

I think another area for speedup could be in loading/caching the partial that gets rendered for each grade. I tried rewriting the views to use collections, but I didn't see a speedup. There's a gem for speeding up partial rendering for collections, but the views became more complicated in order to accommodate collections, so maybe that's for another time.