@pwnall 2016-03-27
Please fix the nits and merge. I'm sure the course staff people will be happier without knowing why :smile: